+++ lib/infra/router/router.dart
@@ -5,31 +5,10 @@ import '../../community/list/communityList.dart';
 import '../../community/list/title/communityListTitle.dart';
 import '../../community/main/communityMain.dart';
 import '../../home/home_screen.dart';
+import '../../home/placeholderPage.dart';
 import '../../home/notice/noticeMain.dart';
 import '../../home/notice/title/notice_title.dart';
 import '../../home/mainpage.dart';
-
-// 임시 페이지 (학식, 대외활동, 취업)
-class PlaceholderPage extends StatelessWidget {
-  final String title;
-
-  const PlaceholderPage({required this.title});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(title),
-      ),
-      body: Center(
-        child: Text(
-          '$title 페이지 준비 중',
-          style: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-        ),
-      ),
-    );
-  }
-}
 
 // GoRouter 설정
 final GoRouter router = GoRouter(
@@ -37,7 +16,7 @@ final GoRouter router = GoRouter(
   routes: [
     ShellRoute(
       builder: (context, state, child) {
-        if (state.fullPath == '/home/notice') {
+        if (state.fullPath == '/notice') {
           return MainPage(
             customTitle: NoticeTitleWidget(), // 커스텀 제목
             customBackgroundColor: Colors.white, // 커스텀 배경색
@@ -64,25 +43,25 @@ final GoRouter router = GoRouter(
             GoRoute(
               path: 'notice',
               pageBuilder: (context, state) => MaterialPage(
-                child: NoticeMain(), // 공지사항 페이지 연결
+                child: NoticeMain(), // 공지사항
               ),
             ),
             GoRoute(
               path: 'cafeteria',
-              pageBuilder: (context, state) => _customTransitionPage(
-                PlaceholderPage(title: '학식'),
+              pageBuilder: (context, state) => MaterialPage(
+                child: PlaceholderPage(title: '학식'), // 학식
               ),
             ),
             GoRoute(
               path: 'activities',
-              pageBuilder: (context, state) => _customTransitionPage(
-                PlaceholderPage(title: '대외활동'),
+              pageBuilder: (context, state) => MaterialPage(
+                child: PlaceholderPage(title: '대외활동'), // 대외활동
               ),
             ),
             GoRoute(
               path: 'employment',
-              pageBuilder: (context, state) => _customTransitionPage(
-                PlaceholderPage(title: '취업'),
+              pageBuilder: (context, state) => MaterialPage(
+                child: PlaceholderPage(title: '취업'), // 취업
               ),
             ),
           ],
